osx/transformer_utils/requirements/runtime.txt
chumpy
dataclasses; python_version == '3.7'
json_tricks
matplotlib
munkres
numpy
opencv-python
pillow
scipy
torchvision
xtcocotools>=1.12
# easydict
# einops

